requirements.txt
rembg==2.0.1
Pillow==9.3.0
gunicorn==20.1.0
azure-functions